+++ nullstone.tf
@@ -1,13 +1,10 @@
-terraform {
-  required_providers {
-    ns = {
-      source  = "nullstone-io/ns"
-      version = "~> 0.8.2"
-    }
-  }
-}
-
 data "ns_workspace" "this" {}
+
+data "ns_agent" "this" {}
+
+locals {
+  ns_agent_user_arn = data.ns_agent.this.aws_user_arn
+}
 
 // Generate a random suffix to ensure uniqueness of resources
 resource "random_string" "resource_suffix" {