+++ secrets.tf
@@ -10,7 +10,7 @@ resource "aws_secretsmanager_secret" "app_secret" {
 
   name                    = "${local.resource_name}/${each.value}"
   tags                    = local.tags
-  kms_key_id              = aws_kms_key.this.id
+  kms_key_id              = module.scaffold.kms_key.id
   recovery_window_in_days = 0 // force delete so that re-adding the secret doesn't cause issues
 
   lifecycle {